refactor: simplify source repository plumbing
