+++ db/app/src/utils/automations.ts
@@ -16,6 +16,7 @@ import {
   createAutomationId,
   createAutomationRunId,
 } from "../schema";
+import { getRowsAffected, isDuplicateKeyError } from "./drizzle-results";
 
 export {
   type AutomationScheduleInput,
@@ -561,41 +562,4 @@ export async function markAutomationRunSkipped(
       )
     );
   return getRowsAffected(result) > 0;
-}
-
-function getRowsAffected(result: unknown): number {
-  if (result === null || typeof result !== "object") {
-    return 0;
-  }
-
-  const { affectedRows, rowsAffected } = result as {
-    affectedRows?: unknown;
-    rowsAffected?: unknown;
-  };
-
-  if (typeof rowsAffected === "number") {
-    return rowsAffected;
-  }
-  if (typeof affectedRows === "number") {
-    return affectedRows;
-  }
-  return 0;
-}
-
-function isDuplicateKeyError(error: unknown): boolean {
-  if (error === null || typeof error !== "object") {
-    return false;
-  }
-
-  const { body, code, message } = error as {
-    body?: { code?: unknown };
-    code?: unknown;
-    message?: unknown;
-  };
-
-  return (
-    body?.code === "ER_DUP_ENTRY" ||
-    code === "ER_DUP_ENTRY" ||
-    (typeof message === "string" && message.includes("Duplicate entry"))
-  );
 }
