+++ packages/github-app-node/src/installation-tokens.ts
@@ -1,14 +1,15 @@
 import { z } from "zod";
 import { GitHubAppNodeError } from "./errors";
+import {
+  fetchGitHubJson,
+  githubJsonHeaders,
+  normalizeGitHubApiBaseUrl,
+} from "./github-api";
 
 const installationTokenResponseSchema = z.object({
   expires_at: z.string().min(1),
   token: z.string().min(1),
 });
-
-function normalizeApiBaseUrl(value: string | undefined) {
-  return (value ?? "https://api.github.com").replace(/\/+$/, "");
-}
 
 export async function createGitHubInstallationToken(input: {
   apiBaseUrl?: string;
@@ -18,31 +19,25 @@ export async function createGitHubInstallationToken(input: {
   installationId: string;
 }): Promise<{ expiresAt: string; token: string }> {
   const requestFetch = input.fetch ?? fetch;
-  const apiBaseUrl = normalizeApiBaseUrl(input.apiBaseUrl);
+  const apiBaseUrl = normalizeGitHubApiBaseUrl(input.apiBaseUrl);
   const url = `${apiBaseUrl}/app/installations/${input.installationId}/access_tokens`;
 
-  let res: Response;
-  try {
-    res = await requestFetch(url, {
+  const { json, response } = await fetchGitHubJson({
+    fetch: requestFetch,
+    init: {
       method: "POST",
-      headers: {
-        accept: "application/vnd.github+json",
-        authorization: `Bearer ${input.appJwt}`,
-        ...(input.apiVersion
-          ? { "x-github-api-version": input.apiVersion }
-          : {}),
-      },
-    });
-  } catch {
-    throw new GitHubAppNodeError(
-      "GITHUB_API_REQUEST_FAILED",
-      "GitHub installation token request failed."
-    );
-  }
+      headers: githubJsonHeaders({
+        apiVersion: input.apiVersion,
+        token: input.appJwt,
+      }),
+    },
+    requestErrorCode: "GITHUB_API_REQUEST_FAILED",
+    requestErrorMessage: "GitHub installation token request failed.",
+    url,
+  });
 
-  const json = await res.json().catch(() => null);
   const parsed = installationTokenResponseSchema.safeParse(json);
-  if (!(res.ok && parsed.success)) {
+  if (!(response.ok && parsed.success)) {
     throw new GitHubAppNodeError(
       "GITHUB_API_RESPONSE_INVALID",
       "GitHub installation token response was invalid."
